Adding 'bbblb' commands to save and restore server and tenant configuration from JSON.
The 'bbblb state save' command exports server and tenant configuration to JSON. The 'bbblb state load' command loads this configuration and applies it to the current database.

This can be used for backup/restore or to sync server state with generated server and tenant configuration files (e.g. via ansible)