change: Tenant cache and join overrides
This patch introduced a short cache for tenant info to improve API latency, and changes the way tenant overrides are persisted and managed to support different override types (e.g. create/join). This required changes in the DB schema and state export format and change or introduced several CLI commands.

closes #5